Change ImageSize to size_t so it can't overflow to a negative value..
Addresses SF tracker bug 1876380, DGifSlurp() tries to allocate negative
amount of memory.  Modern malloc() implementations take size_t so the
whole call chain should be safe.

+++ lib/dgif_lib.c
@@ -1003,7 +1003,7 @@ DGifBufferedInput(GifFileType * GifFile,
 int
 DGifSlurp(GifFileType * GifFile) {
 
-    int ImageSize;
+    size_t ImageSize;
     GifRecordType RecordType;
     SavedImage *sp;
     GifByteType *ExtData;
@@ -1032,7 +1032,7 @@ DGifSlurp(GifFileType * GifFile) {
               if (ImageSize > (SIZE_MAX / sizeof(GifPixelType))) {
                   return GIF_ERROR;
               }
-              sp->RasterBits = (unsigned char *)malloc((size_t)ImageSize *
+              sp->RasterBits = (unsigned char *)malloc(ImageSize *
                       sizeof(GifPixelType));
 
               if (sp->RasterBits == NULL) {
